% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnote.R
\name{backup.notes}
\alias{backup.notes}
\title{Back up your notes to a new location}
\usage{
backup.notes(file = "notes.csv", current_project = FALSE)
}
\arguments{
\item{file}{location to save new file}

\item{current_project}{specify whether to save all notes, or just notes from the current project}
}
\description{
Back up your notes to a new location
}
