% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnote.R
\name{check_project_notes}
\alias{check_project_notes}
\title{Print the most recent notes saved in the current project in a
formatted, readable style}
\usage{
check_project_notes(n = 10, topic = "all")
}
\arguments{
\item{n}{The number of most recent notes to print. Defaults to 10.}

\item{topic}{The topic to filter notes by, default is "all", resulting in no filter.}
}
\description{
This function prints the most recent notes to the console
including the note, date, project of origin and related topics
}
