% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnote.R
\name{checknotes}
\alias{checknotes}
\title{Recall notes you've made with jotdown() By default, only notes from the current project will be recalled, to recall all notes, set current_project = FALSE}
\usage{
checknotes(topic = "all", current_project = TRUE, n = 10)
}
\description{
Recall notes you've made with jotdown() By default, only notes from the current project will be recalled, to recall all notes, set current_project = FALSE
}
