% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnote.R
\name{init_project_notes}
\alias{init_project_notes}
\title{Initialize Project Notes Directory}
\usage{
init_project_notes()
}
\value{
A message indicating whether the project notes directory was created
or if it already exists.
}
\description{
Creates a local `rnotes/` directory in the current project, allowing notes
to be stored and accessed within the project folder. This function is
useful for collaborative work, where team members can share notes related
to a specific project.
}
\details{
If the directory already exists, the function does nothing and informs
the user.
}
\examples{
\dontrun{
init_project_notes()  # Initializes the project notes directory
}
}
