% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnote.R
\name{jotdown}
\alias{jotdown}
\title{Make a note to yourself that you can recall later in the console from any project.}
\usage{
jotdown(note, topics)
}
\arguments{
\item{note}{text content to be stored as a note}

\item{topics}{tag(s) added to the note for filtering with checknotes()}
}
\description{
Make a note to yourself that you can recall later in the console from any project.
}
